requirement.txt
streamlit
python-docx
langchain
langchain-openai
langchain-chroma